exceptions: move UserPermissionDenied to plugin_helpers
Why:

* comply with wazo-notebook/file-structure.md
