+++ wazo_calld/plugins/relocates/services.py
@@ -5,7 +5,6 @@ import re
 import logging
 import threading
 
-from wazo_calld.exceptions import UserPermissionDenied
 from wazo_calld.plugin_helpers import ami
 from wazo_calld.plugin_helpers.ari_ import (
     ARINotFound,
@@ -13,10 +12,11 @@ from wazo_calld.plugin_helpers.ari_ import (
 )
 from wazo_calld.plugin_helpers.confd import User
 from wazo_calld.plugin_helpers.exceptions import (
-    NotEnoughChannels,
-    TooManyChannels,
     InvalidUserLine,
     InvalidUserUUID,
+    NotEnoughChannels,
+    TooManyChannels,
+    UserPermissionDenied,
 )
 
 from .exceptions import (
