+++ wazo_calld/plugins/transfers/services.py
@@ -2,18 +2,21 @@
 # SPDX-License-Identifier: GPL-3.0-or-later
 
 import logging
+
 from ari.exceptions import ARINotFound
 from xivo.caller_id import assemble_caller_id
 
 from wazo_calld.ari_ import DEFAULT_APPLICATION_NAME
-from wazo_calld.exceptions import UserPermissionDenied
 from wazo_calld.exceptions import InvalidExtension
+from wazo_calld.plugin_helpers import ami
 from wazo_calld.plugin_helpers.ari_ import Channel
 from wazo_calld.plugin_helpers.confd import User
-from wazo_calld.plugin_helpers.exceptions import NotEnoughChannels
-from wazo_calld.plugin_helpers.exceptions import TooManyChannels
+from wazo_calld.plugin_helpers.exceptions import (
+    NotEnoughChannels,
+    TooManyChannels,
+    UserPermissionDenied,
+)
 
-from wazo_calld.plugin_helpers import ami
 from . import ari_helpers
 from .exceptions import NoSuchTransfer
 from .exceptions import TooManyTransferredCandidates
